# libs/feature/webhook/swagger.json
{"openapi":"3.0.0","paths":{"/api/webhook/logs":{"get":{"operationId":"WebhookLogsController_findManyLogs","parameters":[{"name":"curPage","required":false,"in":"query","schema":{"type":"number"}},{"name":"perPage","required":false,"in":"query","schema":{"type":"number"}},{"name":"searchText","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"type":"string"}},{"name":"webhookId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindManyWebhookLogResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]}},"/api/webhook/logs/{id}":{"delete":{"operationId":"WebhookLogsController_deleteOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]},"get":{"operationId":"WebhookLogsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebhookLog"}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]}},"/api/webhook/profile":{"get":{"operationId":"WebhookController_profile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebhookUser"}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]}},"/api/webhook/events":{"get":{"operationId":"WebhookController_events","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WebhookEvent"}}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]}},"/api/webhook":{"get":{"operationId":"WebhookController_findMany","parameters":[{"name":"curPage","required":false,"in":"query","schema":{"type":"number"}},{"name":"perPage","required":false,"in":"query","schema":{"type":"number"}},{"name":"searchText","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindManyWebhookResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]},"post":{"operationId":"WebhookController_createOne","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWebhookDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Webhook"}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]}},"/api/webhook/test-request":{"post":{"operationId":"WebhookController_testRequest","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWebhookDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebhookTestRequestResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]}},"/api/webhook/{id}":{"put":{"operationId":"WebhookController_updateOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWebhookDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Webhook"}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]},"delete":{"operationId":"WebhookController_deleteOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]},"get":{"operationId":"WebhookController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Webhook"}}}},"400":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WebhookError"},{"$ref":"#/components/schemas/ValidationError"}]}}}}},"tags":["Webhook"]}}},"info":{"title":"","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"WebhookStatus":{"type":"string","enum":["Pending","Process","Success","Error","Timeout"]},"WebhookRole":{"type":"string","enum":["Admin","User"]},"WebhookUser":{"type":"object","properties":{"id":{"type":"string"},"externalTenantId":{"type":"string"},"externalUserId":{"type":"string"},"userRole":{"allOf":[{"$ref":"#/components/schemas/WebhookRole"}]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"Webhook_Webhook_createdByToWebhookUser":{"type":"array","items":{"$ref":"#/components/schemas/Webhook"}},"Webhook_Webhook_updatedByToWebhookUser":{"type":"array","items":{"$ref":"#/components/schemas/Webhook"}}},"required":["id","externalTenantId","externalUserId","userRole","createdAt","updatedAt"]},"Webhook":{"type":"object","properties":{"id":{"type":"string"},"eventName":{"type":"string"},"endpoint":{"type":"string"},"enabled":{"type":"boolean"},"headers":{"type":"object","nullable":true},"requestTimeout":{"type":"integer","format":"int32","nullable":true},"externalTenantId":{"type":"string"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"workUntilDate":{"type":"string","format":"date-time","nullable":true},"WebhookUser_Webhook_createdByToWebhookUser":{"$ref":"#/components/schemas/WebhookUser"},"WebhookUser_Webhook_updatedByToWebhookUser":{"$ref":"#/components/schemas/WebhookUser"},"WebhookLog":{"type":"array","items":{"$ref":"#/components/schemas/WebhookLog"}}},"required":["id","eventName","endpoint","enabled","headers","requestTimeout","externalTenantId","createdBy","updatedBy","createdAt","updatedAt","workUntilDate"]},"WebhookLog":{"type":"object","properties":{"id":{"type":"string"},"request":{"type":"object"},"responseStatus":{"type":"string"},"response":{"type":"object","nullable":true},"webhookStatus":{"allOf":[{"$ref":"#/components/schemas/WebhookStatus"}]},"webhookId":{"type":"string"},"externalTenantId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"Webhook":{"$ref":"#/components/schemas/Webhook"}},"required":["id","request","responseStatus","response","webhookStatus","webhookId","externalTenantId","createdAt","updatedAt"]},"FindManyResponseMeta":{"type":"object","properties":{"curPage":{"type":"number"},"perPage":{"type":"number"},"totalResults":{"type":"number"}},"required":["totalResults"]},"FindManyWebhookLogResponse":{"type":"object","properties":{"webhookLogs":{"type":"array","items":{"$ref":"#/components/schemas/WebhookLog"}},"meta":{"$ref":"#/components/schemas/FindManyResponseMeta"}},"required":["webhookLogs","meta"]},"StatusResponse":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"WebhookEvent":{"type":"object","properties":{"eventName":{"type":"string"},"description":{"type":"string"},"example":{"type":"object"}},"required":["eventName","description","example"]},"FindManyWebhookResponse":{"type":"object","properties":{"webhooks":{"type":"array","items":{"$ref":"#/components/schemas/Webhook"}},"meta":{"$ref":"#/components/schemas/FindManyResponseMeta"}},"required":["webhooks","meta"]},"CreateWebhookDto":{"type":"object","properties":{"eventName":{"type":"string"},"endpoint":{"type":"string"},"enabled":{"type":"boolean"},"headers":{"type":"object","nullable":true},"requestTimeout":{"type":"integer","format":"int32","nullable":true},"workUntilDate":{"type":"string","format":"date-time","nullable":true}},"required":["eventName","endpoint","enabled"]},"WebhookTestRequestResponse":{"type":"object","properties":{"request":{"type":"object"},"responseStatus":{"type":"string"},"response":{"type":"object","nullable":true},"webhookStatus":{"allOf":[{"$ref":"#/components/schemas/WebhookStatus"}]}},"required":["request","responseStatus","response","webhookStatus"]},"UpdateWebhookDto":{"type":"object","properties":{"eventName":{"type":"string"},"endpoint":{"type":"string"},"enabled":{"type":"boolean"},"headers":{"type":"object","nullable":true},"requestTimeout":{"type":"integer","format":"int32","nullable":true},"workUntilDate":{"type":"string","format":"date-time","nullable":true}}},"WebhookErrorEnum":{"type":"string","enum":["WEBHOOK-000","WEBHOOK-001","WEBHOOK-002","WEBHOOK-003","WEBHOOK-004","WEBHOOK-005"]},"WebhookError":{"type":"object","properties":{"message":{"type":"string","description":"Webhook error (WEBHOOK-000), Tenant ID not set (WEBHOOK-003), User ID not set (WEBHOOK-002), Forbidden (WEBHOOK-001), User not found (WEBHOOK-004), Event not found (WEBHOOK-005)","example":"Webhook error"},"code":{"example":"WEBHOOK-000","allOf":[{"$ref":"#/components/schemas/WebhookErrorEnum"}]},"metadata":{"type":"object"}},"required":["message","code"]},"WebhookScalarFieldEnum":{"type":"string","enum":["id","eventName","endpoint","enabled","headers","requestTimeout","externalTenantId","createdBy","updatedBy","createdAt","updatedAt","workUntilDate"]},"WebhookLogScalarFieldEnum":{"type":"string","enum":["id","request","responseStatus","response","webhookStatus","webhookId","externalTenantId","createdAt","updatedAt"]},"WebhookUserScalarFieldEnum":{"type":"string","enum":["id","externalTenantId","externalUserId","userRole","createdAt","updatedAt"]},"WebhookEntities":{"type":"object","properties":{"Webhook":{"allOf":[{"$ref":"#/components/schemas/WebhookScalarFieldEnum"}]},"WebhookLog":{"allOf":[{"$ref":"#/components/schemas/WebhookLogScalarFieldEnum"}]},"WebhookUser":{"allOf":[{"$ref":"#/components/schemas/WebhookUserScalarFieldEnum"}]}},"required":["Webhook","WebhookLog","WebhookUser"]}}}}